ticketbuddy_new/requirements.txt
asgiref==3.11.0
awsebcli==3.25.3
backports.tarfile==1.2.0
bcrypt==5.0.0
blessed==1.25.0
boto3==1.41.5
botocore==1.41.5
build==1.3.0
cement==2.10.14
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
decorator==5.2.1
Deprecated==1.3.1
Django==4.2.26
docutils==0.22.3
fabric==3.2.2
gunicorn==23.0.0
id==1.5.0
idna==3.11
importlib_metadata==8.7.0
invoke==2.2.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jeepney==0.9.0
jmespath==1.0.1
keyring==25.7.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.8.0
nh3==0.3.2
packaging==24.2
paramiko==4.0.0
pathspec==0.12.1
pillow==11.3.0
pycparser==2.23
Pygments==2.19.2
PyNaCl==1.6.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
PyYAML==6.0.3
qrcode==8.2
readme_renderer==44.0
reportlab==4.4.5
requests==2.32.5
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.2.0
s3transfer==0.15.0
SecretStorage==3.3.3
semantic-version==2.10.0
six==1.17.0
sqlparse==0.5.4
termcolor==2.5.0
ticketdiscount==0.1.0
tomli==2.3.0
twine==6.2.0
typing_extensions==4.15.0
urllib3==1.26.20
wcwidth==0.2.14
wrapt==2.0.1
zipp==3.23.0
